mvp.v2.1 commit

+++ src/main/java/com/example/summerpractic2023/repos/UserRepo.java
@@ -1,8 +1,10 @@
 package com.example.summerpractic2023.repos;
 
 import com.example.summerpractic2023.base.entity.User;
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo  extends JpaRepository<User, Long> {
+public interface UserRepo extends JpaRepository<User, Integer> {
     User findByUsername(String username);
+
 }
